pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "data-agents-copilot"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.30.0",
    "mcp>=1.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "questionary>=2.0",
    "rich>=13.0.0",
    "chainlit>=1.1.0",
    "requests>=2.31.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "bandit>=1.7.0",
    "pre-commit>=3.7.0",
    "pytest>=8.0.0",
    "pytest-cov>=5.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
data-agent = "cli.main:main"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=agents --cov=config --cov=hooks --cov-report=term-missing --cov-fail-under=80"

[tool.setuptools]
packages = ["agents", "cli", "config", "hooks", "memory", "workflow", "evals", "ui", "utils", "orchestrator"]
